change pydantic to dataclass in domain

+++ src/app/interfaces/schemas.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, Field, EmailStr
@@ -5,20 +6,19 @@ from typing import Optional, Annotated
 
 from regex import regex
 
-
 # API-контракты
 
 
 class ItemCreate(BaseModel):
     name: Annotated[str, Field(max_length=127), "Maximal length of name is 127 symbols."]
-    price: float
-    tax: Optional[float]
+    price: Decimal
+    tax: Optional[Decimal]
     description: Annotated[
         Optional[str], Field(max_length=1023), "Maximal length of description is 1023 symbols."
     ] = None
-    price: Annotated[float, Field(gt=-1e9, lt=1e10), "Price is between -10e9 and 10e15."]
+    price: Annotated[Decimal, Field(gt=-1e9, lt=1e10), "Price is between -10e9 and 10e15."]
     tax: Annotated[
-        Optional[float], Field(gt=-100, lt=1e5), "Tax is between -100 and 10e6 percentage."
+        Optional[Decimal], Field(gt=-100, lt=1e5), "Tax is between -100 and 10e6 percentage."
     ] = None
